Move class to MathUtils and rename to "Angles"
