+++ osu.Framework/Graphics/Containers/BufferedContainer_DrawNode.cs
@@ -12,7 +12,6 @@ using System;
 using osu.Framework.Graphics.Colour;
 using osu.Framework.MathUtils;
 using osuTK.Graphics.ES30;
-using osu.Framework.Extensions.NumericExtensions;
 
 namespace osu.Framework.Graphics.Containers
 {
@@ -108,7 +107,7 @@ namespace osu.Framework.Graphics.Containers
                     Vector2 size = current.Size;
                     blurShader.GetUniform<Vector2>(@"g_TexSize").UpdateValue(ref size);
 
-                    float radians = -NumericExtensions.DegreesToRadians(blurRotation);
+                    float radians = -Angles.DegreesToRadians(blurRotation);
                     Vector2 blur = new Vector2(MathF.Cos(radians), MathF.Sin(radians));
                     blurShader.GetUniform<Vector2>(@"g_BlurDirection").UpdateValue(ref blur);
